backend/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.1
